issue: 2123641 Fix Nginx root user permissions run
We have now forbidden flow in fork creation for Nginx workload
for root user.

The forbidden flow is as follows:
In Nginx: Master process forks new worker process --->
In VMA fork call interception: We reset parent contexts
and reopen them for child process --->
In Nginx: In case root user used, Nginx will set GID,
UID and user name for the user chosen in the conf file --->
In VMA: Any call to verbs from now on will be blocked,
because of permission changes done in Nginx calls
changing GID, UID and user name.

The temporary solution, until RSS implementation done:
The solution for regular user and root user is separated
and done in different locations. It is forbidden to
change UID, GID and username of the process after the call
to ibv_open_device done in VMA context reopening.
Thus, for regular user, it will be done in fork call, since
in this case, Nginx will not call to setuid function.
For root user, it will be done in setuid call interception
by VMA, after the call to the OS setuid function.

Signed-off-by: Simon Raviv <simonra@mellanox.com>