issue: 2140026 sendfile: Fix VMA error during RO file mapping
VMA prints an error when it tries to do RW open of RO file.
This can occur on sendfile flow.
In this case VMA instead of oppening a file with RW is calling dup,
and is doing private mapping of the file instead of shared mapping.
This commits reduces log level from error to debug, since VMA WA the issue.

Signed-off-by: Alex Vainman <alexv@mellanox.com>