issue: 2342345 Workaround for long waiters list
If multiple epfds watch for an event on a single fd, it leads to a
linear search within kernel on an event. With 64K sockets, we have
64K epfds watching a single cq_fd. As results CPU usage is 100% in
kernel space on the CPU core which handles CQ interrupts. This is
a bottleneck and waste of resources.

This commit is a dirty workaround. It works only in case of single CQ.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>