issue: 3115378 Print verbose info on error CQE
When a completion reports an error CQ mgr prints info from the WC
element. However, it lacks important fields: syndrome and hw codes.
Using the codes we can determine the reason of the error completion.

Print missed info in cq_mgr_mlx5 implementation where we have access
to the codes.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>