issue: 2647722 Reduce TCP segments batching
TCP sockets request free TCP segments by 64. This is a hardcoded value.
Therefore, to open N TCP connections we need at least (N * 64) allocated
segments.

To serve 350K connections, we have to allocate at least ~2.6GB of
segment objects. This is an issue for a system with low memory, because
we also have to allocate TX/RX/ZC buffers and socket objects.

Reduce number of segments requested at a time to 4. With this change we
need to allocate at least 170MB of segments what is more acceptable
memory footprint.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>