issue: 2648131 Force 5T rules for outgoing TCP sockets
3T rules are redundant for outgoing sockets and lead to an issue in
the NGINX scenario.

The issue is that NGINX can create outgoing connections in proxy mode
and the sockets are not distinguished between listen, incoming and
outgoing. Therefore, a 3T rule is created for such an outgoing
connection and NGINX-related code applies a mask to the rule to match
a range of source ports. Such a rule may not match the real remote port.

As solution, force 5T rules for all outgoing TCP sockets.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>
Signed-off-by: Alex Briskin <abriskin@nvidia.com>