issue: 2575576 Add ZCOPY send operation using user key
This case can be used with sendmsg(SCM_VMA_PD) when the
data portion contains an uint32_t array of the memory keys
based on the PD. Number of elements in this
array should be equal to msg_iovlen value. Every data pointer
in msg_iov has correspondent memory key.

Signed-off-by: Igor Ivanov <igori@nvidia.com>