issue: 2047225 Don't bind to the same port in adjacent tests
Commit d1dc7c8 changes close() behaviour and shadow socket is closed in
socket destructor now. But destructor is called in the internal thread
what creates a race window between calling close() and actual socket
destruction. If application tries to bind to the same address within
this window it receives EADDRINUSE error.

To avoid the above issue, don't call bind to the same address in
adjacent tests.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>