issue: 2647722 Fix corner cases for TLS blocking socket
Don't restrict sending data by available sndbuf for blocking sockets.
We will block in sockinfo_tcp::tcp_tx() in this case.

Don't retry if g_b_exit flag is set, because it leads to an endless
loop.

Improve error handling when TCP tx() fails in the middle of a TLS record
for a blocking socket.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>