issue: 2670696 Destroy pending sockets on exit
TCP sockets may not close immediately. If the application terminates
before such sockets transit to CLOSED state within TCP timer, they are
lost and not destroyed. As result, statistics is not dumped to the
VMA_STATS_FILE what is done in the socket destructor.

Return the pending list to fd_collection to keep track of such sockets.
The list was removed as part of commit d1dc7c8.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>