issue: 2672303 Fix sendfile zerocopy TX flow
When sndbuf doesn't allow to send entire IOV element with a single
tcp_write() call, it is possible that on the next iteration we will
be able to send more data after tx_wait(). In this case we have to
advance tx_ptr pointer to write new data on the next loop iteration.

The issue is that for sendfile implementation (TX_FILE operation) we
advance not tx_ptr, but "file_offset". When zerocopy flag is set we have
to use tx_ptr.

This issue also affects TLS TX flow which reuses sendfile zerocopy flow.
As result, we can send corrupted data.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>