issue: 2679474 Fix segfault during termination phase
During termination phase g_b_exit flag is set to true and
sockinfo_tcp::tcp_tx() returns with EINTR error. The issue is that some
data can be queued as TCP segments. In this case number of queued bytes
must be returned.

Because of the above issue, TLS code thinks that no data is queued and
destroys TLS record. And when the queued pbufs are freed, they tries to
access the destroyed record.

As solution, put reference to failed records instead of destroying it.
We can do this, because tls_record doesn't do any additional logic when
reference counting drops to 0.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>