issue: 2680431 Fallback to SW TLS when TIS/DEK fails
Currently there is assert() on that TIS and DEK objects must be created
successfully. This is not correct behavior and it leads to unexpected
segfault in release mode when happens.

Instead, check for TIS/DEK return value and status of obtaining TIS
number. When any step fails return an error, os openssl falls back to
software TLS implementation.

This patch handles error path, but doesn't address root cause of TIS/DEK
creation failure.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>