issue: 2685998 Don't mask nor create 2nd rule for UDP
In NGINX related code we apply a mask to 3T rules depending on number of
NGINX worker processes and also create a 2nd rule in configuration of
non power of 2 workers. This affects UDP sockets since we create 3T
rules for UDP.

Avoid NGINX related manipulations with rules for UDP. This will allow to
properly work with outgoing UDP traffic in context of NGINX application.
This also fixes error messages that are related to mistaken 2nd hardware
UDP rule when a UDP socket is closed by NGINX.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>