issue: 2686808 Don't set TCP PSH flag for throughput profiles
When XLIO sends a TSO segment with the PSH flag, card sets the flag only
to the last split segment. Linux kernel with GRO functionality can
receive GRO segments out of order for such a TCP stream. This leads to
large number of fast retransmits.

This is not optimal behavior of Linux kernel and we can mitigate this by
removing PSH flag for outgoing segments in throughout profiles.

This patch reduces unnecessary retransmits in throughput scenarios.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>