issue: 2716793 Handle SYN-ACK on TIME-WAIT sockets
If a SYN-ACK is received on a TIME-WAIT socket, it is handled as an
initial SYN. This scenario can happen for an outgoing connection if
remote host retransmits SYN-ACK while application closes socket in
ESTABLISHED state.

In the above scenario we initiate an incoming connection request and
reuse the outgoing socket like it was an incoming.

To resolve the issue, differentiate between SYN and SYN-ACK. Also send
RST in case of SYN-ACK on TIME-WAIT socket since we should do so for out
of state SYN-ACK.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>