issue: 2621445 Create TIR object in RX ring
Proper RQ belongs to RX ring which may be different than TX ring.
Therefore, we must create TIR objects in the RX ring to point to the
right RQ.

Created TIR objects are still cached in the TX ring, since this is
required for reference counting. TX lock is required to protect TIR
cache.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>