issue: 2824476 Fix for unsafe m_pendig_to_remove_lst processing
It is not guaranteed that the same sockinfo object is met once
in this loop that processes sockinfo timer events.
So in case sockinfo object is destroyed other processing
of the same object mast be ingored.
Note: Check on is_cleaned() is not safe completely. More stable solution
should be considered.

Signed-off-by: Igor Ivanov <igori@nvidia.com>