issue: 2621445 Make a plain list of buffers in TLS RX
GRO can build a pbuf chain that represents a single TCP segment. LwIP
layer then pushes this chain to the socket layer where it is inserted
into a list as a single element. Therefore, we build a 2-level list:

  buf ----------------> buf --------------------> ...
   |                     |
   +-> pbuf -> pbuf      +-> pbuf -> pbuf-> ...

This is not convenient to have such a list and work with it as a plain
data stream. To optimize performance and to refactor TLS RX flow,
Split GRO pbuf chain into individual buffers and insert them into a
plain list.

Current commit changes only TLS socket layer and doesn't change TCP
socket layer.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>