issue: 2621445 Reset TLS layer on connection close
TLS layer destructor returns TIS and TIR objects to the TX ring. If we
do this in context of sockinfo_tcp destructor, we will lock TX ring
in the internal thread for a long time, because DEK object destruction
is a long operation.

To avoid lock contention, reset TLS layer on connection close, so we
do this in the main thread.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>