issue: 2842211 Split rexmit segment into single-pbuf segments
This avoids the scenario when a pbuf in the middle of a TCP segment can
be freed during a spurious retransmission. We track reference counting
only for first pbufs in TCP segments.

When this happens, such a ZC pbuf can free the underlying object such as
TLS record or sendfile mapping. But since we still have the spurious
retransmission in the send queue, we can send corrupted data due to
use-after-free.

For TCP protocol this is not a big issue, since corrupted spuriously
retransmitted segments will be thrown away as duplicates. However, the
corrupted data can go through TLS context on NIC and break TLS offload.
So new segments will be sent as non-offloaded.

With split retransmitted segments we guarantee that spurious
retransmissions will always hold references to the underlying objects
like TLS records, even if the original TCP segment has a pbuf in the
middle that points to different underlying object than the 1st pbuf.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>