issue: 2667588 Support IPV6_V6ONLY sockopt for listen all
This commit adds copy_sockopt_fork to copy the IPV6_ONLY flag for forked sockets.
This socket options copy is currently implemented for nginx and for very specific options.
This copy is called as part of fork() flow of nginx specifically.
If a generic fork() is implemented, this copy should be reimplemented in a more generic way,
see is_inherited_option mechanism of sockinfo_tcp for an example.

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>