issue: 3010282 Reduce resolve timeout if no link-local IPv6
IPv6 address resolution is expected to fail if there is no a link-local
address on the outgoing interface. The root cause is in Linux kernel
which searches for link-local source address in case of RDMA address
resolution command.

Therefore, there is no point in long timeout. Reduce this timeout to
minimal 1ms for failure scenario.

We cannot avoid rdma_resolve_addr() call, because kernel drops
unexpected NS replies otherwise. Combination of the failed RDMA command
and manual NS request makes kernel to accept the reply and put it to the
neigh table.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>