issue: 2783472 Recreating rdma channel in case of rdma_create_id failure
On RH8.3 it is observed that rdma_create_id fails because the channel
which was created before security context switches of the process does not
pass security checks. Recreating the channel at the stage of
rdma_create_id failure, creates the channel with appropriate security context
and second rdma_create_id try passes successfully.

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>