issue: 2972491 Remove global tcp_input_pcb
tcp_input_pcb is used as a global variable to check whether tcp_output()
is called from the RX path. In case of handling RX/TX paths for
different sockets in different threads, this can lead to ignoring of
tcp_output() in the TX path.

Also handling of RX on different rings concurrently can lead to
unprotected concurrent modification of the variable.

As a solution, use per socket flag instead of global variable.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>