issue: 2885753 Analysis comments for socket reuse in process of closing
In scenario of high CPS, in sockinfo_tcp::prepare_to_close we call tcp_close()
which may send a FIN, and then we call handle_socket_linger().
handle_socket_linger may receive a FIN, move the connection to TIME_WAIT state
and immediately receive a SYN, which will cause the connection to initiate
a reuse sequence while we are still in process of closing it (in prepare_to_close).
This breaks internal structures completely.
To overcome this issue, we should postpone the ability of the socket to be reusable
until the closing sequence is complete.

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>