issue: 3113450 Reset ZC buffer's ctx pointer on close()
Because of delayed TX completion, zerocopy notification ballback can
be called after of in the middle of sockinfo object destruction.
This leads to a segfault, because the callback tries to insert an
element to the error queue in the socket.

To resolve the race, reset tx.zc.ctx pointer in close() path, so
subsequent TX completions ignore the callback.

This makes impact only on sockets with setsockopt(SO_ZEROCOPY). The
cost is a loop over inflight WQEs under TX ring lock.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>