issue: 3139461 Fix listen() on accepted socket
Incoming (accepted) socket object is in TCP_SOCK_BOUND state.
prepareListen() doesn't handle an incoming socket and treats such
a socket as a valid created and bound socket. This leads to a
listen(2) syscall on the shadow socket.

With the incoming shadow socket removal feature, the listen(2)
syscall now fails with ENOTSOCK error.

Check for the above scenario explicitly and return EINVAL as expected.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>