issue: 2667588 Use if-index as net-dev-entry registration key
Today net-dev-entry is used to track bonding related changes per interface.
However, the registration key is ip_addr which is translated to if_index
by IP lookup. In case of many IPs the same interface will be registered
to notifications, multiple time, what is a waste.

Since net-dev-entry tracks nothing related to IP,
the key for net-dev-entry should be the interface index.
This prevents a lot of redundant translations between IP and an interface,
and also prevents multiple identical registration per interface.

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>