issue: 3139461 Handle ioctl and fcntl if no shadow sockfd
When there is no shadow socket for an incoming TCP connection, we cannot
call kernel's syscalls on the fake fd. Therefore, we have to handle
ioctl() and fcntl() calls.

Return ENOTSUP error if an option is not supported by XLIO and it's not
handled in other way. Any required option must be implemented in XLIO
explicitly.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>