issue: 2667588 IPv6 Fixing routing source address decisions
The previous behaviour checked if routing entry has src-addr, and
if it didn't, then for that routing-entry the src-addr was forecfully
set by looking up the address on the coresponding device.
Later this routing-src-addr was used as src-addr of the packet.
This is not suitable for IPv6. For IPv6 routing-src-addr is used
only if it was explicitly configured for the entry.
Otherwise, the IPv6 address selection algorithm decides which src addr to use.

Signed-off-by: Alexander Grissik <agrissik@nvidia.com>