issue: 3097181 Add release of multiple buffer refs under a single lock
SSL_sendfile allocates a TX buffer and uses it for multiple TLS record
headers. To avoid races, TLS socket takes references for all the blocks
in the buffer and on the socket destruction it must free extra refs.
Again, to avoid races, this needs to be done under ring TX lock, so
this change introduces new API which releases multiple references under
a single lock.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>