issue: 2842211 Don't leave unacked empty on send failure
If a send operation fails (for example, due to lack of TX WQEs) we don't
move the respective TCP segment to the unacked queue. The idea of this is
to resend the failed segment on the next tcp_output() and avoid a
retransmission or RTO. However, if the unacked queue remains empty it can
affect RTO and the socket will stop progressing.

To avoid the issue, keep a failed segment as unsent only if the unacked
queue is not empty. This will guarantee that RTO will trigger
retransmissions in case there are no other triggers.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>