issue: 2842211 Retransmit dropped unacked segment
We don't move a dropped segment back to unsent queue if the unacked
queue is empty. Otherwise, this would break RTO in multiple places such
as slow timer, tcp_rexmit_rto() and during handling of incoming
segments.

However, this can lead to RTO in a scenario of continuous send syscalls
while it could be avoided. For example, this happens when the 1st
segment is dropped, moved to the unacked queue, but the 2nd segment is
too big to fit into the TCP window. The 2nd segment cannot be split due
to the condition and the TCP stream is stuck until RTO happens.
In this case, a blocking socket is blocked when the send buffer becomes
full.

To avoid unnecessary RTO in the above case, mark the 1st dropped, so
it can be forcibly retransmitted on the next tcp_output() call.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>