issue: 3202977 Fix type overflow during trimming TCP seg
With LRO/GRO TCP segments can be large and 16bit type is not enough to
fit possible offsets/lengths. This can lead to a type overflow and
broken pbuf chains as result. Further, a broken chain breaks accounting
of received data what leads to a warning or even a segfault.

Increase type length for pbuf_header() and the trimming code. This fixes
pbuf chain. Also increase type length for pbuf_realloc() to avoid
similar issues in the future.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>