issue: 3293037 Fix clang-analyzer warning in sock-redirect
The warning claims that the global fd_collection object can be
dereferenced when it's NULL. To avoid this, move the code block under
existing condition (g_p_fd_collection && offload_sockets). The 2nd part
of the condition also applies for the code block, because this is
necessary condition to change add_to_udp_pool to true.

Also refactor related code.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>