issue: 3319895 Poll until the last TX CQE
Since WQEs are completed in order, it is enough to get the last
completion and handle all the previous indexes. This approach also
avoids growing distance between ci and pi in some scenarios.

Don't skip error completions. We need to log them and accumulate in
the statistics.

Refactor poll_and_process_element_tx() to remove copy-paste.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>