issue: 3388776 Improve systemd check for chroot environment
See: http://0pointer.de/blog/projects/changing-roots

First of all, systemctl detects when it is run in a chroot.
If so, most of its operations will become NOPs, with the exception
of systemctl enable and systemctl disable. If a package installation
script hence calls these two commands, services will be enabled
in the guest OS. However, should a package installation script
include a command like systemctl restart as part of the package
upgrade process this will have no effect at all when run
in a chroot() environment.

Signed-off-by: Igor Ivanov <igori@nvidia.com>