issue: 3433756 Delete attach flow data object with proper type
RFS class casts IPv4/IPv6 specific objects to a C structure and then deletes
it in the destructor. This leads to a type mismatch between allocation and
destruction.

This doesn't look like a real issue, because the types don't have specific
destructors and the allocator doesn't use type size during deallocation.
However, analyzers (such as ASAN) generate a warning.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>