issue: 3502107 Fix missed sockets in the stats file
During XLIO finalization, internal thread is stopped and all unfinished
actions are dropped eventually. Also, there is a race window during
which new actions are dropped.

Don't skip socket destruction in the internal thread if the thread is
stopped. This is a corner case of the last event loop iteration. This
resolves issue with missed socket statistics partially, however, in
most cases this solution is enough.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>