issue: 3455924 Don't treat ethernet padding as payload
In the flow_id case mem_buf_desc_t::sz_data contains ethernet padding.
Further, this padding is treated as TCP payload in 2 specific flows.

The root cause of the issue is that in some places payload is calculated
as sz_data minus header size. Remove padding like it is done for
non-flow_id case, so sz_data is always real PDU size.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>